+++ components/StatusIcon/vue/StatusIcon.rootstory.tsx
@@ -2,25 +2,25 @@ import clsx from 'clsx'
 import StatusIcon from './StatusIcon.vue'
 import { statuses } from '../constants'
 
-const sizes = ['4', '8', '12', '16', '24']
+const sizes = ['4', '8', '12', '16', '24'] as const
 
 export default () => (
-  <table className="w-full">
+  <table class="w-full">
     <thead>
       <tr key="h0">
-        <th className="text-left">Status</th>
-        <th className="text-left">Variant</th>
-        <th className="text-left" colSpan={sizes.length}>
+        <th class="text-left">Status</th>
+        <th class="text-left">Variant</th>
+        <th class="text-left" col-span={sizes.length}>
           Size
         </th>
-        <th className="text-left">Used for</th>
+        <th class="text-left">Used for</th>
       </tr>
       <tr key="h1">
         <th />
         <th />
         {sizes.map((size) => {
           return (
-            <th key={`heading-${size}`} className="text-left">
+            <th key={`heading-${size}`} class="text-left">
               {size}
             </th>
           )
@@ -30,14 +30,14 @@ export default () => (
     </thead>
     <tbody>
       {Object.keys(statuses).map((status) => {
-        const statusInfo = statuses[status]
+        const statusInfo = statuses[status as keyof typeof statuses]
 
         return (
           <>
             {statusInfo.variants.map((variant, i) => {
               return (
                 <tr
-                  className={clsx(i === 0 && 'border-t')}
+                  class={clsx(i === 0 && 'border-t')}
                   key={`${status}-${variant}`}
                 >
                   {i === 0 ? (
@@ -45,7 +45,7 @@ export default () => (
                       {statusInfo.link ? (
                         <a
                           href={statusInfo.link}
-                          className="text-indigo-500 underline"
+                          class="text-indigo-500 underline"
                           target="_blank"
                         >
                           {status}
@@ -61,16 +61,16 @@ export default () => (
                   <td>{variant}</td>
                   {sizes.map((size) => {
                     return (
-                      <td key={`${status}-${size}-${variant}`} className="py-2">
+                      <td key={`${status}-${size}-${variant}`} class="py-2">
                         <StatusIcon
-                          status={status}
+                          status={status as keyof typeof statuses}
                           size={size}
                           variant={variant}
                         />
                       </td>
                     )
                   })}
-                  <td className="py-2 align-top">
+                  <td class="py-2 align-top">
                     {i === 0 ? statusInfo.use : ''}
                   </td>
                 </tr>