+++ components/Tooltip/react/TooltipReact.cy.tsx
@@ -7,8 +7,8 @@ import assertions from '../assertions'
 
 describe('Tooltip', { viewportHeight: 800, viewportWidth: 800 }, () => {
   function mountStory(options: Parameters<typeof TooltipStory>[0] = {}) {
-    mount(<TooltipStory {...options} tabIndex={1} />)
+    mount(<TooltipStory {...options} tabindex={1} />)
   }
 
-  assertions(mountStory)
+  assertions(mountStory, 'react')
 })