+++ css/src/index.ts
@@ -1,24 +1,3 @@
-import type { UserOptions } from 'vite-plugin-windicss'
-import VitePlugin from 'vite-plugin-windicss'
-import WebpackPlugin from 'windicss-webpack-plugin'
-import { getConfig } from './get-config'
-
-export const CyCSSVitePlugin = (options: UserOptions = {}) => {
-  const VitePluginCJS = VitePlugin as any
-  if (typeof VitePluginCJS.default === 'function') {
-    return VitePluginCJS.default(getConfig(options))
-  }
-  return VitePlugin(getConfig(options))
-}
-
-export const CyCSSWebpackPlugin = (options: UserOptions) => {
-  const WebpackPluginCJS = WebpackPlugin as any
-  if (typeof WebpackPluginCJS.default === 'function') {
-    return WebpackPluginCJS.default(getConfig(options))
-  }
-  return new WebpackPlugin(getConfig(options))
-}
-
 export * from './colors'
 
 export {
@@ -26,8 +5,5 @@ export {
   ADDITIONAL_COLORS,
 } from './icon-extractor-tools'
 
-export { default as WindKeepRollupPlugin } from './wind-keep-rollup-plugin'
 export { default as TailwindConfig } from './tailwind.config'
 export { IconExtractor as TailwindIconExtractor } from './tw-icon-extractor'
-export { IconExtractor as WindiIconExtractor } from './wc-icon-extractor'
-export { default as WindiConfig } from './windi.config'