+++ docs/.vitepress/theme/components/Layout.vue
@@ -1,6 +1,6 @@
 <script lang="ts" setup>
-import './markdown.scss'
-import './fonts/fonts.css'
+import '../assets/markdown.scss'
+import '../assets/fonts/fonts.css'
 import { useData, useRouter } from 'vitepress'
 import {
   computed,
@@ -132,7 +132,7 @@ const { frontmatter } = useData() as any
   <template v-if="!frontmatter.layout">
     <DarkModeSwitch class="fixed z-50 bottom-[32px] right-[32px]" />
     <header
-      class="flex flex-row-reverse md:flex-row fixed z-40 w-full bg-white dark:bg-gray-1000 h-[72px] justify-between items-center px-[24px] md:px-[32px] border-b border-gray-100 dark:border-gray-50/07"
+      class="flex flex-row-reverse md:flex-row fixed z-40 w-full bg-white dark:bg-gray-1000 h-[72px] justify-between items-center px-[24px] md:px-[32px] border-b border-gray-100 dark:border-gray-50/[.07]"
     >
       <a href="/">
         <img
@@ -147,7 +147,7 @@ const { frontmatter } = useData() as any
       <Button
         href="https://github.com/cypress-io/cypress-design"
         variant="outline-light"
-        class="fixed bottom-[34px] right-[116px] md:static text-indigo-500 !bg-gray-50 !dark:bg-gray-900 !rounded-full !dark:border-gray-50/07 !dark:hover:border-gray-50/20 !dark:hover:shadow-gray-50/07"
+        class="fixed bottom-[34px] right-[116px] md:static text-indigo-500 !bg-gray-50 !dark:bg-gray-900 !rounded-full !dark:border-gray-50/[.07] !dark:hover:border-gray-50/20 !dark:hover:shadow-gray-50/07"
       >
         <IconSocialGithubSolid />
       </Button>
@@ -164,7 +164,7 @@ const { frontmatter } = useData() as any
       />
       <div class="w-[250px] hidden md:block shrink-0" />
       <aside
-        class="fixed py-[32px] bg-white dark:bg-gray-1000 z-50 transition transition-transform duration-300 h-[calc(100vh-72px)] overflow-auto border-r border-gray-100 dark:border-gray-50/07"
+        class="fixed py-[32px] bg-white dark:bg-gray-1000 z-50 transition transition-transform duration-300 h-[calc(100vh-72px)] overflow-auto border-r border-gray-100 dark:border-gray-50/[.07]"
         :class="{
           'transform -translate-x-full md:translate-x-0': !mobileMenuOpen,
         }"