+++ packages/eslint-plugin/lib/rules/deprecate-imports-again.js
@@ -1,0 +1,5 @@
+/**
+ * This duplicate allows to use the same rule twice with different options
+ */
+
+module.exports = require('./deprecate-imports')