+++ test/vue-app/src/App.vue
@@ -5,8 +5,10 @@ import Checkbox from '@cypress-design/vue-checkbox'
 import Icon, {
   IconActionQuestionMarkCircle,
   IconObjectBookCode,
+  IconObjectMagnifyingGlass,
 } from '@cypress-design/vue-icon'
 import Alert from '@cypress-design/vue-alert'
+// @ts-expect-error see this issue https://github.com/vuejs/language-tools/issues/1662
 import Spinner from '@cypress-design/vue-spinner/sfc'
 import Button from '@cypress-design/vue-button'
 
@@ -14,11 +16,12 @@ const checked = ref(false)
 </script>
 
 <template>
-  <div style="width: 500px; text-align: center; margin: auto" class="group">
+  <div class="group w-[500px] text-center mx-auto">
     <Checkbox v-model="checked" label="Checkbox" id="Sample" />
     <p v-if="checked">Checked</p>
     <IconObjectBookCode
       size="48"
+      class="icon-dark-red-300"
       :stroke-color="checked ? 'indigo-600' : 'red-300'"
       fill-color="indigo-300"
       hover-fill-color="red-300"
@@ -36,11 +39,16 @@ const checked = ref(false)
     <Accordion title="hi">
       hello
       <template #iconEl>
-        <IconActionQuestionMarkCircle
-          strokeColor="jade-600"
-          fillColor="jade-50"
-        />
+        <IconActionQuestionMarkCircle fillColor="jade-50" />
       </template>
     </Accordion>
+    <div class="group p-4 bg-jade-50">
+      <IconObjectMagnifyingGlass
+        interactive-colors-on-group
+        strokeColor="jade-600"
+        fillColor="jade-50"
+        focus-within-stroke-color="red-500"
+      /><input />
+    </div>
   </div>
 </template>