+++ css/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cypress-design/css
 
+## 0.13.3
+
+### Patch Changes
+
+- [#244](https://github.com/cypress-io/cypress-design/pull/244) [`2b8927e`](https://github.com/cypress-io/cypress-design/commit/2b8927e6b6ee395d63a1b12e11c57888ee43940e) Thanks [@elevatebart](https://github.com/elevatebart)! - remove windi loader
+
+- [#245](https://github.com/cypress-io/cypress-design/pull/245) [`d52d023`](https://github.com/cypress-io/cypress-design/commit/d52d02301bb851d514661a8c258d0c4ae5baf171) Thanks [@elevatebart](https://github.com/elevatebart)! - use tailwind to rollup classes in the final package
+
+## 0.13.2
+
+### Patch Changes
+
+- [#219](https://github.com/cypress-io/cypress-design/pull/219) [`0a4ce3b`](https://github.com/cypress-io/cypress-design/commit/0a4ce3bb55b9ab0510f7fb9ba207d5fdc3cb723f) Thanks [@elevatebart](https://github.com/elevatebart)! - refactor some typings so eslint passes without warning
+
 ## 0.13.1
 
 ### Patch Changes